Text improvement 1. Text - Layout text & icons. * Feature added to layout text and icons.   Icons can be added using the mark-up string feature.
  i.e.
   "Hello <item 'url'='icon.png'/>"

  The icon can be resized with the 'width' and 'height' properties.

  "Hello <item 'url'='icon.png' 'width'=24 'height'=24/>"

  Optionally only a size can be specified. The layout engine will layout the text with the gaps of the given size and
  retrieve the layout info of those gaps. It's the responsability of the user to fill the pixel buffer with the images.

* New 'item' tag added to the mark-up processor.

2. Text - Bitmap fonts
* Add support for bitmap fonts.

3. Text renderer utils - Set color to glyphs.
* Required to set the color of the embedded items
  and the glyphs of the bitmap fonts.
* Attribute added to the <item> tag that
  specifies if the color of the item has to be
  blended with the color of the text.
* <color> tag processed and sent to the
  TextRenderer class in adaptor.

* The color of font and bitmap font glyphs can be set
  individually with the <color> tag in a markup string
  or globally if the default color is set.
  If no default color is set, font glyphs will be white
  and bitmap font glyphs will use their colors.

* By default embedded items will use their colors. If
  the attribute 'color-blending' is set to 'multiply'
  the color of the text will be multiplied by the color
  of the item image. If the image is white the result
  will be the color of the text.

  "Hello <item 'url'='image.png' 'color-blending'='multiply'>

4. Text Utils - Render color Emojis.

* Cairo renders Emojis in black and white. This
  patch tells the Cairo renderer whether each
  glyph is an Emoji.

5. Circular text implementation.
* New parameters added to the DevelText::RendererParameters
  struct for circular text.
  - layout: The type of layout: one of {"singleLine",
    "multiLine", "circular"}
  - circularAlignment: The text alignment within the arc:
    one of {"begin", "center", "end"}
  - radius: The radius in pixels of the circular text.
    Measured from the center to the top of the line.
  - beginAngle: The begin angle in degrees of the text area
    on the circle. The top of the circle is 0°, the right
    side 90°, the bottom 180° and the left 270°.
  - incrementAngle: The increment angle in degrees of the
    text area on the circle. The @p incrementAngle defines
    a direction. If positive, the text will be laid out
    clockwise.

  @note the horizontalAlignment and verticalAlignment can
  be used to align the whole text within the text area.

* Example: How to set parameters for circular text from
  a C++ application

  Dali::Toolkit::DevelText::RendererParameters textParameters;
  textParameters.text = "Hello world, this is a demo of circular text.";
  textParameters.horizontalAlignment = "center";
  textParameters.verticalAlignment = "center";
  textParameters.circularAlignment = "center";
  textParameters.layout = "circular";
  textParameters.textColor = Color::BLACK;
  textParameters.fontSize = 16.f;
  textParameters.textWidth = 360u;
  textParameters.textHeight = 360u;
  textParameters.radius = 130u;
  textParameters.beginAngle = 180.f;
  textParameters.incrementAngle = 360.f;
  textParameters.ellipsisEnabled = true;

6. Text Utils - Rotate emojis/embedded items.
* Retrieves the layout for rotated embedded items
  added in a second step.
* When the emoji glyph is the one that is replaced
  by the ellipsis glyph, the 'isEmoji' vector pased
  to the cairo-renderer needs to be updated to
  state that '...' is not an emoji.

7. Software italic/bold issues fixed.
* Fixes issues with the synthesis of bold and italic style by software.

+++ dali-toolkit/internal/text/text-model.cpp
@@ -162,7 +162,7 @@ const Vector4& Model::GetOutlineColor() const
   return mVisualModel->GetOutlineColor();
 }
 
-float Model::GetOutlineWidth() const
+uint16_t Model::GetOutlineWidth() const
 {
   return mVisualModel->GetOutlineWidth();
 }
